Remove indice se existir

+++ db/migrate/20191004150447_add_discarded_at_to_daily_frequency_student_index.rb
@@ -1,6 +1,12 @@
 class AddDiscardedAtToDailyFrequencyStudentIndex < ActiveRecord::Migration
   def change
-    remove_index :daily_frequency_students, name: 'daily_frequency_students_daily_frequency_id_student_id_idx'
+    if index_exists?(
+      :daily_frequency_students,
+      [:daily_frequency_id, :student_id],
+      name: 'daily_frequency_students_daily_frequency_id_student_id_idx'
+    )
+      remove_index :daily_frequency_students, name: 'daily_frequency_students_daily_frequency_id_student_id_idx'
+    end
 
     add_index(
       :daily_frequency_students,
